+++ apps/wave-frontend/src-tauri/.gitignore
@@ -1,4 +1,3 @@
 # Generated by Cargo
 # will have compiled files and executables
 /target/
-
