+++ apps/wave-frontend/src-tauri/build.rs
@@ -1,3 +1,3 @@
 fn main() {
-    tauri_build::build()
+  tauri_build::build()
 }
