+++ apps/wave-frontend/src/lib/api.ts
@@ -1,0 +1,6 @@
+import { edenTreaty } from "@elysiajs/eden";
+import type { App } from "wave-backend";
+
+const api = edenTreaty<App>("http://0.0.0.0:3000");
+
+export default api;
