+++ apps/wave-frontend/svelte.config.js
@@ -1,7 +1,18 @@
-import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
+import adapter from "@sveltejs/adapter-static";
+import { vitePreprocess } from "@sveltejs/kit/vite";
 
-export default {
-    // Consult https://svelte.dev/docs#compile-time-svelte-preprocess
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+    // Consult https://kit.svelte.dev/docs/integrations#preprocessors
     // for more information about preprocessors
     preprocess: vitePreprocess(),
+
+    kit: {
+        // adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+        // If your environment is not supported or you settled on a specific environment, switch out the adapter.
+        // See https://kit.svelte.dev/docs/adapters for more information about adapters.
+        adapter: adapter(),
+    },
 };
+
+export default config;
