--- apps/wave-frontend/tailwind.config.ts
@@ -1,14 +1,0 @@
-import type { Config } from "tailwindcss";
-
-import tailwindForms from "@tailwindcss/forms";
-
-const config = {
-    darkMode: "class",
-    content: ["./src/**/*.{html,js,svelte,ts}"],
-    theme: {
-        extend: {},
-    },
-    plugins: [tailwindForms()],
-} satisfies Config;
-
-export default config;
