--- apps/wave-frontend/turbo.json
@@ -1,9 +1,0 @@
-{
-  "extends": ["//"],
-  "pipeline": {
-    "web": {
-      "cache": false,
-      "persistent": true
-    }
-  }
-}
